pyproject.toml
[tool.poetry]
name = "datawarehouses"
version = "0.1.0"
description = ""
authors = ["vlad.luca <vlad.luca@syneto.eu>"]

[tool.poetry.dependencies]
python = "^3.8"
cassandra-driver = "^3.25.0"
python-dotenv = "^0.20.0"
SQLAlchemy = "^1.4.34"
black = "^22.3.0"
pyspark = "^3.2.1"
pandas = "^1.4.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
